# Add info about source code location in JSON logging