@serathius 2021-05-30
Thanks, can you add tests?