@MadhavJivrajani 2021-05-31
Hey @serathius, thanks for reminding me about writing tests. I have a small question:

Since in implementations like `Info`, an `Entry` is used directly, and in that since the `Caller` field is not specified, even by mentioning the `EncodeCaller`, `zap` will ignore including information about source code. What I tried doing was something similar to the way `zap.NewProduction()` does it:
```go
func (l *zapLogger) Info(msg string, keysAndVals ...interface{}) {
    checkedEntry := l.l.Check(zap.InfoLevel, msg)
    checkedEntry.Write(l.handleFields(keysAndVals)...)
}
```
but the issue with this is that the source code information that it logs corresponds to the line: `checkedEntry := l.l.Check(zap.InfoLevel, msg)`. The logger returned by `NewProduction()` works as intended. I'm not sure what I'm missing, any pointers?