@serathius 2021-06-04
To add caller information to logs you need to do three things:
* Set "EncodeCaller" on encoderConfig field as you did.
* Set "CallerKey" on encoderConfig
* Set "Caller" field on entries in Info, Error and dPanic fields.

Code to set "Caller" field (taken from zap internals)
```
	frame, defined := getCallerFrame(1)
	if defined {
		entry.Caller = zapcore.EntryCaller{
			Defined:  defined,
			PC:       frame.PC,
			File:     frame.File,
			Line:     frame.Line,
			Function: frame.Function,
		}
	}
```