@serathius 2021-06-05
Please add a release note as this PR definitely does a user facing change.