@serathius 2021-06-05
One additional request, could you run benchmarks to get performance cost that we are paying to put caller information. This may influence our decision if this should be enabled by default or not.