@serathius 2021-06-05
ok, I found out that running benchmarks in K8s is pretty hard. Let me just put results that I got.
```
name              old time/op    new time/op    delta
InfoLoggerInfo-8     925ns ± 3%    1618ns ± 9%  +74.98%  (p=0.000 n=10+8)
ZapLoggerError-8    1.22µs ±11%    1.91µs ±10%  +56.63%  (p=0.000 n=9+10)
ZapLoggerV-8        1.02µs ± 2%    1.93µs ±11%  +90.29%  (p=0.000 n=8+10)

name              old alloc/op   new alloc/op   delta
InfoLoggerInfo-8    1.15kB ± 0%    1.40kB ± 0%  +21.60%  (p=0.000 n=9+10)
ZapLoggerError-8    1.32kB ± 0%    1.57kB ± 0%  +18.81%  (p=0.000 n=10+10)
ZapLoggerV-8        1.17kB ± 0%    1.42kB ± 0%  +21.26%  (p=0.000 n=10+10)

name              old allocs/op  new allocs/op  delta
InfoLoggerInfo-8      4.00 ± 0%      7.00 ± 0%  +75.00%  (p=0.000 n=10+10)
ZapLoggerError-8      7.00 ± 0%     10.00 ± 0%  +42.86%  (p=0.000 n=10+10)
ZapLoggerV-8          5.00 ± 0%      8.00 ± 0%  +60.00%  (p=0.000 n=10+10)
```