@MadhavJivrajani 2021-06-05
Thank you @serathius 😄 
For future reference, how did you run the benchmark? 

Also, I've added the release note.