@serathius 2021-06-05
I didn't find any ready command so I have edited ./hack/make-rules/test.sh file