@serathius 2021-06-05
Hey, I started looking into integration into klog and found that we are missing things to make caller work not only as a logr interface, but also with klog. We are missing some integration that would show this problem. Let me create separate PR that will add nesesery tests.