@serathius 2021-06-05
@MadhavJivrajani tests were merged please rebase, and check the errors in new tests.
I found that when error wraps logr it requires it to also implement CallDepthLogSink interface.
https://github.com/go-logr/logr/blob/606acd2ddfcb374e4ee6c5ee8741e1a7689d5927/logr.go#L367-L374