@MadhavJivrajani 2021-06-19
/hold cancel
Hi @serathius, I've made the changes and implemented the `CallDepthLogger` interface. The integration tests helped a lot, thanks! PTAL and lmk what you think and feedback if any :)